Power Manager panel refined Integrated Battery panel into Power Manager Panel Network Management panel refined
