+++ lib/ripple/platform/tizen/2.0/spec/ui.js
@@ -20,7 +20,6 @@ module.exports = {
         "geoView",
         "widgetConfig",
         "deviceSettings",
-        "battery",
         "application",
         "network",
         "power",
