--- lib/ripple/ui/plugins/battery/panel.html
@@ -1,87 +1,0 @@
-<!--
- *  Copyright 2011 Intel Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
--->
-<section id="battery-container" class="ui-box ui-state-default">
-    <section class="h2 info-header">
-        <section class="collapse-handle"><span class="ui-box-TitleImage ui-box-TitleImageClosed"></span>Battery</section>
-        <section class="drag-handle">
-            <span><img src="images/menuDraggerIcon.png"/></span>
-        </section>
-    </section>
-
-    <section class="info ui-widget-content ui-corner-all" style="display: none;">
-        <table class="panel-table">
-            <tr>
-                <td>
-                    <span class="ui-text-label" id="charging-state">Plugged in?</span>
-                </td>
-                <td>
-                    <input id="is-charging" type="checkbox"/>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <label class="ui-text-label">Battery Level:</label>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <input id="battery-volume" type="number" value="100" min="0" max="100" step="1" class="ui-state-default ui-corner-all">
-                    <label id="battery-volume-label" style="display:none"></label>
-                    <label>%</label>
-                </td>
-                <td>
-                    <div id="block0" class="battery-block"></div>
-                    <div id="block1" class="battery-block"></div>
-                    <div id="block2" class="battery-block"></div>
-                    <div id="block3" class="battery-block"></div>
-                    <div id="block4" class="battery-block"></div>
-                    <div id="block5" class="battery-block"></div>
-                </td>
-            </tr>
-            <tr>
-                <td colspan="2">
-                    <label class="ui-text-label" id="remaining-time-label">Remaining Time (seconds)</label>
-                </td>
-            </tr>
-        </table>
-        <div>
-            <label id="remain-time"></label>
-        </div>
-
-        <div align="center">
-            <button id="simulate-low-battery-btn">Simulate low battery</button>
-            <button id="stop-btn">Stop</button>
-        </div>
-        <div id="simulate-low-battery" title="Simulate low battery">
-            <p style="margin-left:5px">
-                <label>Start at :&nbsp;&nbsp;</label>
-                <input id="low-battery-start-at" type="number" value="20" min="0" max="100" step="1" class="ui-state-default ui-corner-all">
-                <label>%</label>
-            </p>
-            <p style="margin-left:5px">
-                <label>Drop to 0% over :&nbsp;&nbsp;</label>
-                <input id="low-battery-remain-time" value="1200">
-                <select>
-                    <option value="sec">sec</option>
-                </select>
-            </p>
-            <p align="right">
-                <button id="simulate-low-battery-cancel">Cancel</button>
-                <button id="simulate-low-battery-start">Start</button>
-            </p>
-        </div>
-    </section>
-</section>
