+++ lib/ripple/ui/plugins/network.js
@@ -142,14 +142,39 @@ function _initializeElements() {
         $("#nfc-radio").buttonset();
         $("#bluetooth-radio").buttonset();
 
+        $("#wifi-radio1").prop('checked', false);
+        $("#wifi-radio2").prop('checked', false);
+        if (deviceSettings.retrieve("WifiNetwork.status") === "true") {
+            $("#wifi-radio1").prop('checked', true);
+        }
+        else {
+            $("#wifi-radio2").prop('checked', true);
+        }
+        $("#wifi-radio1").button("refresh");
+        $("#wifi-radio2").button("refresh");
+
+        $("#network-type1").prop('checked', false);
+        $("#network-type2").prop('checked', false);
+        $("#network-type3").prop('checked', false);
+        if (deviceSettings.retrieve("Network.networkType") === "2G") {
+            $("#network-type1").prop('checked', true);
+        }
+        else if (deviceSettings.retrieve("Network.networkType") === "3G") {
+            $("#network-type2").prop('checked', true);
+        }
+        else {
+            $("#network-type3").prop('checked', true);
+        }
+        $("#network-type1").button("refresh");
+        $("#network-type2").button("refresh");
+        $("#network-type3").button("refresh");
+
         $("#wifi-radio1").click(function () {
-            $("#wifi-main-container").show();
             deviceSettings.persist("WifiNetwork.status", "true");
             event.trigger("WiFiNetworkStatusChanged", [true]);
         });
 
         $("#wifi-radio2").click(function () {
-            $("#wifi-main-container").hide();
             deviceSettings.persist("WifiNetwork.status", "false");
             event.trigger("WiFiNetworkStatusChanged", [false]);
         });
