+++ lib/ripple/ui/plugins/network/panel.html
@@ -28,74 +28,41 @@
                 <td><label class="ui-text-label">WiFi</label></td>
                 <td>
                 <div id="wifi-radio">
-                    <input type="radio" id="wifi-radio1" name="radio" />
+                    <input type="radio" id="wifi-radio1" name="radio-wifi" />
                     <label for="wifi-radio1">On</label>
-                    <input type="radio" id="wifi-radio2" name="radio" checked="checked" />
+                    <input type="radio" id="wifi-radio2" name="radio-wifi" />
                     <label for="wifi-radio2">Off</label>
                 </div></td>
             </tr>
         </table>
-        <div id="wifi-main-container" style="display:none; background-color:#D3D1D1;">
-            <table class="tf_panel-table2">
-                <tr>
-                    <td><span class="ui-text-label">SSID</span></td>
-                    <td>
-                    <input type="text" value="Tizen WiFi" id="device-settings-WifiNetwork-ssid" class="ui-state-default ui-corner-all">
-                    </td>
-                </tr>
-                <tr>
-                    <td><span class="ui-text-label">IP Address</span></td>
-                    <td>
-                    <input type="text" value="192.168.0.1" id="device-settings-WifiNetwork-ipAddress" class="ui-state-default ui-corner-all">
-                    </td>
-                </tr>
-                <tr>
-                    <td><span class="ui-text-label">Signal Strength</span></td>
-                    <td>
-                    <select id="device-settings-WifiNetwork-signalStrength" class="ui-state-default ui-corner-all">
-                        <option value="0">0</option>
-                        <option value="1">1</option>
-                        <option value="2">2</option>
-                        <option value="3">3</option>
-                        <option value="4">4</option>
-                        <option value="5">5</option>
-                        <option value="6">6</option>
-                        <option value="7">7</option>
-                        <option value="8">8</option>
-                        <option value="9">9</option>
-                        <option value="10">10</option>
-                    </select></td>
-                </tr>
-            </table>
-        </div>
         <table class="panel-table">
             <tr>
-                <td colspan="2">&nbsp;---------------------------------------------------------------------------</td>
+                <td colspan="2"><hr></td>
             </tr>
             <tr>
                 <td><label class="ui-text-label">Network Data Type</label></td>
                 <td>
                 <div id="network-type-radio">
-                    <input type="radio" id="network-type1" name="radio" />
+                    <input type="radio" id="network-type1" name="radio-network-data" />
                     <label for="network-type1">2G</label>
-                    <input type="radio" id="network-type2" name="radio" />
+                    <input type="radio" id="network-type2" name="radio-network-data" />
                     <label for="network-type2">3G</label>
-                    <input type="radio" id="network-type3" name="radio" checked="checked" />
+                    <input type="radio" id="network-type3" name="radio-network-data" />
                     <label for="network-type3">4G</label>
                 </div></td>
             </tr>
         </table>
         <table class="panel-table">
             <tr>
-                <td colspan="2">&nbsp;---------------------------------------------------------------------------</td>
+                <td colspan="2"><hr></td>
             </tr>
             <tr>
                 <td><label class="ui-text-label">NFC</label></td>
                 <td>
                 <div id="nfc-radio">
-                    <input type="radio" id="nfc-radio1" name="radio" />
+                    <input type="radio" id="nfc-radio1" name="radio-NFC" />
                     <label for="nfc-radio1">On</label>
-                    <input type="radio" id="nfc-radio2" name="radio" checked="checked" />
+                    <input type="radio" id="nfc-radio2" name="radio-NFC" checked="checked" />
                     <label for="nfc-radio2">Off</label>
                 </div></td>
             </tr>
@@ -341,15 +308,15 @@
         </div>
         <table class="panel-table">
             <tr>
-                <td colspan="2">&nbsp;---------------------------------------------------------------------------</td>
+                <td colspan="2"><hr></td>
             </tr>
             <tr>
                 <td><label class="ui-text-label">Bluetooth</label></td>
                 <td>
                 <div id="bluetooth-radio">
-                    <input type="radio" id="bluetooth-radio1" name="radio" />
+                    <input type="radio" id="bluetooth-radio1" name="radio-bt" />
                     <label for="bluetooth-radio1">On</label>
-                    <input type="radio" id="bluetooth-radio2" name="radio" checked="checked" />
+                    <input type="radio" id="bluetooth-radio2" name="radio-bt" checked="checked" />
                     <label for="bluetooth-radio2">Off</label>
                 </div></td>
             </tr>
@@ -381,7 +348,7 @@
 
             <table>
                 <tr>
-                    <td colspan="2">&nbsp;---------------------------------------------------------------------------
+                    <td colspan="2"><hr>
                     </td>
                 </tr>
             </table>
